+++ plane.h
@@ -27,7 +27,7 @@ class Plane: public Surface {
 public:
 	Plane(const Material &material, const Vector normal, const double distance) : Surface(material), normal(normal), distance(distance) {}
 	virtual std::string toString() const;
-	virtual int intersect(Ray &r, float &t);
+	virtual int intersect(Ray &r, double &t);
 	virtual Vector normalAt(Vector& point) const;
 private:
 	const Vector normal;
