+++ sphere.h
@@ -31,7 +31,7 @@ public:
 		this->radius = radius;
 	}
 	virtual std::string toString() const;
-	virtual int intersect(Ray &r, float &t);
+	virtual int intersect(Ray &r, double &t);
 	virtual Vector normalAt(Vector& point) const;
 private:
 	Vector position;
